sous/server/requirements.txt
fastapi
uvicorn
youtube_transcript_api
bs4
groq
dotenv
chromadb
websockets
deepgram-sdk